% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.r
\name{add_list}
\alias{add_list}
\title{Create a new list. Create a Smart List if a filter is provided.}
\usage{
add_list(name, filter = NULL)
}
\arguments{
\item{name}{the list name}

\item{filter}{criteria for a Smart List}
}
\value{
respose from RTM server
}
\description{
Create a new list. Create a Smart List if a filter is provided.
}

