% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_funs.r
\name{add_task}
\alias{add_task}
\title{Add a task.}
\usage{
add_task(name)
}
\arguments{
\item{name}{the name of the task including all Smart Add tags}
}
\value{
response from the server
}
\description{
Add a task.
}

