% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{addTags}
\alias{addTags}
\title{Calls corresponding RTM task method}
\usage{
addTags(task, tags)
}
\arguments{
\item{task}{the task}

\item{tags}{tags}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

