% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{addTask}
\alias{addTask}
\title{Add a task.}
\usage{
addTask(name)
}
\arguments{
\item{name}{the name of the task including all Smart Add tags}
}
\value{
response from the server
}
\description{
Add a task.
}

