% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{aggregate_task_column}
\alias{aggregate_task_column}
\title{Handle the task column properly depending on its form, either
a data frame or a list.}
\usage{
aggregate_task_column(task_col)
}
\arguments{
\item{task_col}{the task "column", never actually an atomic vector though}
}
\value{
a data frame
}
\description{
If it's a data frame do nothing. If it's a list, it's a list of named
lists whose names are those of the data frame's columns, so converting
is not too difficult. Typically each list has one element but typically
at least one list will have more than one element (which is why it
couldn't have been stored as a data frame in the first place).
}

