% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{apply_rtm_method}
\alias{apply_rtm_method}
\title{Apply a method to a task.}
\usage{
apply_rtm_method(method, task, msg = NULL, ...)
}
\arguments{
\item{method}{one of the RTM API "rtm.tasks.*" methods}

\item{task}{a task}

\item{msg}{the additional parameter that some methods optionally have}

\item{...}{additional arguments}
}
\value{
response from the server
}
\description{
Apply a method to a task.
}

