% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_dev.r
\name{assign_task_method_funs}
\alias{assign_task_method_funs}
\title{Assign}
\usage{
assign_task_method_funs(method_args = task_method_args(),
  fun_factory = make_rtm_task_fun, envir = globalenv())
}
\arguments{
\item{method_args}{a named list of the methods and their arguments}

\item{fun_factory}{the function factory to generate task_method functions}

\item{envir}{environment to evaluate in}
}
\description{
Assign
}

