% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.r
\name{cbind_quickdf}
\alias{cbind_quickdf}
\title{Takes an element x of length 1 and a list that can be quickdf'd,
and returns a cbinded df.}
\usage{
cbind_quickdf(x, l, x_name)
}
\arguments{
\item{x}{an vector of length 1}

\item{l}{a list that can be quickdf'd}

\item{x_name}{the name of the x column}
}
\value{
a data frame
}
\description{
This should be faster than most other possibilities, but of course
it does not check inputs.
}

