% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/response.r
\name{check_http_response}
\alias{check_http_response}
\title{Check the HTTP server response, throw an error if needed.}
\usage{
check_http_response(rsp)
}
\arguments{
\item{rsp}{object returned from \code{httr::GET()}}
}
\description{
Check the HTTP server response, throw an error if needed.
}

