% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.r
\name{check_json_response}
\alias{check_json_response}
\title{Check the JSON sent by the server.}
\usage{
check_json_response(rsp)
}
\arguments{
\item{rsp}{text presumably in JSON format}
}
\description{
Check the JSON sent by the server.
}

