% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/auth.r
\name{get_auth}
\alias{get_auth}
\title{Authenticate a user with the RTM API.}
\usage{
get_auth(perms = "delete")
}
\arguments{
\item{perms}{the permissions the user will have, in increasing order
"read", "write", "delete"}
}
\value{
an authorization token if successful
}
\description{
See https://www.rememberthemilk.com/services/api/authentication.rtm
for details.
}

