% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lists.r
\name{get_rtm_list_id}
\alias{get_rtm_list_id}
\title{Get a list id from a name.}
\usage{
get_rtm_list_id(name)
}
\arguments{
\item{name}{a list name}
}
\value{
a list id as a character string
}
\description{
Get a list id from a name.
}

