% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lists.r
\name{get_rtm_list_name}
\alias{get_rtm_list_name}
\title{Get the RTM list name from an id.}
\usage{
get_rtm_list_name(id)
}
\arguments{
\item{id}{the RTM list id}
}
\value{
the list name
}
\description{
Get the RTM list name from an id.
}

