% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.r
\name{handle_multiple_tasks}
\alias{handle_multiple_tasks}
\title{Handle the case where a task is repeating and has multiple entries
per row.}
\usage{
handle_multiple_tasks(df)
}
\arguments{
\item{df}{the list}
}
\value{
the data frame
}
\description{
In this case df[["task"]] is a list and needs to be converted to a
data frame and merged back into the larger data frame.
}

