% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmauth_init.r
\name{init_rtmauth}
\alias{init_rtmauth}
\title{Retrieve RTM access token.}
\usage{
init_rtmauth(endpoint, app, permission = "read",
  is_interactive = interactive(), private_key = NULL)
}
\arguments{
\item{endpoint}{The RTM endpoint, created by 'rtm_endpoint'}

\item{app}{The RTM consumer application, created by 'rtm_app'}

\item{permission}{optional, a string of permissions to ask for}

\item{is_interactive}{Is the current environment interactive?}

\item{private_key}{not used}
}
\value{
the RTM json content response
}
\description{
See:
https://www.rememberthemilk.com/services/api/authentication.rtm,
which is similar to the OAuth authentication scheme. This code
hews closely to httr::init_oauth1.0, to integrate well with that
package's Token class features.
}

