% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{lookup_task_method_args}
\alias{lookup_task_method_args}
\title{Look up the argument that a method takes}
\usage{
lookup_task_method_args(method)
}
\arguments{
\item{method}{the method as a character string}
}
\value{
the argument as a character string
}
\description{
Look up the argument that a method takes
}

