% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_dev.r
\name{make_rtm_task_fun}
\alias{make_rtm_task_fun}
\title{A closure to create functions corresponding to "rtm.tasks.*"
methods that call apply_rtm_method}
\usage{
make_rtm_task_fun(method, args)
}
\arguments{
\item{method}{the RTM task method}

\item{args}{the names of the arguments to the RTM task method}
}
\value{
a function that calls "rtm_method" with the specified method
}
\description{
A closure to create functions corresponding to "rtm.tasks.*"
methods that call apply_rtm_method
}

