% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{move_priority}
\alias{move_priority}
\title{Calls corresponding RTM task method}
\usage{
move_priority(task, direction)
}
\arguments{
\item{task}{the task}

\item{direction}{direction}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

