% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{move_to}
\alias{move_to}
\title{Move a task from one list to another}
\usage{
move_to(task, to_list_name)
}
\arguments{
\item{task}{the task}

\item{to_list_name}{the name of the list to move it to}
}
\value{
response from the server
}
\description{
Move a task from one list to another
}

