% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{notes.add}
\alias{notes.add}
\title{Calls corresponding RTM task method}
\usage{
notes.add(task, note_title, note_text)
}
\arguments{
\item{task}{the task}

\item{note_title}{note title}

\item{note_text}{note text}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

