% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_funs.r
\name{notes_add}
\alias{notes_add}
\title{Calls corresponding RTM task method}
\usage{
notes_add(task, note_title, note_text)
}
\arguments{
\item{task}{the task}

\item{note_title}{note title}

\item{note_text}{note text}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

