% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlogram_rtm.R
\name{plot.rtm_correlogram}
\alias{plot.rtm_correlogram}
\title{Plot a residual spatial correlogram}
\usage{
\method{plot}{rtm_correlogram}(x, ...)
}
\arguments{
\item{x}{An \code{rtm_correlogram} object from \code{\link[=correlogram_rtm]{correlogram_rtm()}}.}

\item{...}{Passed on to \code{\link[=plot]{plot()}}.}
}
\description{
Plot a residual spatial correlogram
}
