% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_rtm.R
\name{plot.rtm_validation}
\alias{plot.rtm_validation}
\title{Plot the ROC curve of a risk terrain model validation}
\usage{
\method{plot}{rtm_validation}(x, ...)
}
\arguments{
\item{x}{An \code{rtm_validation} object from \code{\link[=validate_rtm]{validate_rtm()}}.}

\item{...}{Passed on to \code{\link[=plot]{plot()}}.}
}
\description{
Plot the ROC curve of a risk terrain model validation
}
