% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{rbinddf}
\alias{rbinddf}
\title{Create a data frame out of two lists or data frames.}
\usage{
rbinddf(x, y)
}
\arguments{
\item{x}{a list or data frame}

\item{y}{a list or data frame}
}
\value{
a data frame
}
\description{
Create a data frame out of two lists or data frames.
}

