% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/note.r
\name{rtm_add_note}
\alias{rtm_add_note}
\title{Add a note to a task}
\usage{
rtm_add_note(task, title, text)
}
\arguments{
\item{task}{the task}

\item{title}{the title of the note}

\item{text}{the text of the note}
}
\value{
the response
}
\description{
Add a note to a task
}

