% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{rtm_add_task}
\alias{rtm_add_task}
\title{Add a task.}
\usage{
rtm_add_task(name)
}
\arguments{
\item{name}{the name of the task including all Smart Add tags}
}
\value{
response from the server
}
\description{
Add a task.
}

