% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lists.r
\name{rtm_cache_lists}
\alias{rtm_cache_lists}
\title{Cache the list of RTM lists so we only need it once per R session.}
\usage{
rtm_cache_lists(varname = ".rtm_lists_info", envir = globalenv())
}
\arguments{
\item{varname}{the name the variable is to be stored in in the global
environment}

\item{envir}{environment to store it in, should be the global}
}
\description{
Cache the list of RTM lists so we only need it once per R session.
}

