% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.r
\name{rtm_cache_timeline}
\alias{rtm_cache_timeline}
\title{Request a new timeline and store it in the global environment.}
\usage{
rtm_cache_timeline(varname = ".rtm_timeline", envir = globalenv())
}
\arguments{
\item{varname}{the name of the global variable to store it in}

\item{envir}{the environment, by default the global environment}
}
\description{
Request a new timeline and store it in the global environment.
}

