% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{rtm_complete}
\alias{rtm_complete}
\title{Complete a task.}
\usage{
rtm_complete(task)
}
\arguments{
\item{task}{a single row task from a data frame of tasks that is the
result from rtm_search}
}
\value{
response from the server
}
\description{
Complete a task.
}

