% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.r
\name{rtm_date}
\alias{rtm_date}
\title{Convert date stored in RTM server format to R POSIXlt format.}
\usage{
rtm_date(date)
}
\arguments{
\item{date}{date string in RTM server format}
}
\value{
a POSIXlt value
}
\description{
Convert date stored in RTM server format to R POSIXlt format.
}

