% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{rtm_move_task}
\alias{rtm_move_task}
\title{Move a task from one list to another}
\usage{
rtm_move_task(task, to_list_name)
}
\arguments{
\item{task}{the task}

\item{to_list_name}{the name of the list to move it to}
}
\value{
response from the server
}
\description{
Move a task from one list to another
}

