% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{rtm_pat}
\alias{rtm_pat}
\title{Retrieves the personal authentication token}
\usage{
rtm_pat(force = FALSE)
}
\arguments{
\item{force}{force entry of new PAT}
}
\value{
a character string
}
\description{
Retrieves the personal authentication token
}

