% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{rtm_req}
\alias{rtm_req}
\title{Make an HTTP GET request of the RTM service.}
\usage{
rtm_req(method, ..., simpliflat = TRUE)
}
\arguments{
\item{method}{the RTM method}

\item{...}{additional parameters}

\item{simpliflat}{use jsonlite simplify and flatten options}
}
\value{
a list or data frame, as parsed by jsonlite::fromJSON
}
\description{
Make an HTTP GET request of the RTM service.
}

