% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search.r
\name{rtm_search}
\alias{rtm_search}
\title{Search tasks using the RTM advanced search feature.}
\usage{
rtm_search(filter)
}
\arguments{
\item{filter}{the search terms passed to RTM advanced search}
}
\value{
a list of data frames contained in the response from rtm_req
    for an advanced search
}
\description{
See https://www.rememberthemilk.com/help/?ctx=basics.search.advanced
for specific parameters.
}
\examples{
rtm_search("list:Shopping")

rtm_search("priority:1 list:Work")
}

