% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.r
\name{rtm_task_method}
\alias{rtm_task_method}
\title{Call rtm_req with a "rtm.tasks.*" method on the given task.}
\usage{
rtm_task_method(method, task, ...)
}
\arguments{
\item{method}{one of the RTM API "rtm.tasks.*" methods}

\item{task}{a task}

\item{...}{additional arguments}
}
\value{
response from the server
}
\description{
Call rtm_req with a "rtm.tasks.*" method on the given task.
}

