% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.r
\name{rtm_time}
\alias{rtm_time}
\title{Convert the time from text to UTC via the RTM service.}
\usage{
rtm_time(text)
}
\arguments{
\item{text}{a character string representing text}
}
\value{
a time
}
\description{
Convert the time from text to UTC via the RTM service.
}

