% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timeline.r
\name{rtm_timeline}
\alias{rtm_timeline}
\title{Get the timeline in use or request a new one.}
\usage{
rtm_timeline()
}
\value{
a timeline id character string
}
\description{
Get the timeline in use or request a new one.
}

