% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmauth_init.r
\name{rtmauth_listener}
\alias{rtmauth_listener}
\title{Open a browser and wait for user to indicate authorization.}
\usage{
rtmauth_listener(request_url, is_interactive = interactive())
}
\arguments{
\item{request_url}{the url to send to the browser}

\item{is_interactive}{is an interactive environment available}
}
\value{
user entry value
}
\description{
Open a browser and wait for user to indicate authorization.
}

