% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmauth_signature.r
\name{rtmauth_signature}
\alias{rtmauth_signature}
\title{Sign parameters for an RTM request.}
\usage{
rtmauth_signature(url, method = "GET", app, token = NULL, frob = NULL,
  ...)
}
\arguments{
\item{url}{Url of request}

\item{method}{method of request}

\item{app}{RTM app object}

\item{token}{token}

\item{frob}{frob}

\item{...}{additional parameters}
}
\value{
list object of length one named 'api_sig' containing the signature
}
\description{
See "Signing Requests" at:
https://www.rememberthemilk.com/services/api/authentication.rtm
}
\details{
Follows the design of httr::oauth_signature.
}

