% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmauth_token.r
\name{rtmauth_token}
\alias{rtmauth_token}
\title{Main function to call to get an RTM token.}
\usage{
rtmauth_token(endpoint, app, permission = "read", private_key = NULL,
  cache = getOption("httr_oauth_cache"))
}
\arguments{
\item{endpoint}{The RTM endpoint}

\item{app}{the RTM app}

\item{permission}{a string of permissions to ask for}

\item{private_key}{not used}

\item{cache}{uses the cache functionality of the Token class}
}
\value{
A TokenRTM reference class object.
}
\description{
Modeled after oauth1.0_token, uses the httr::Token class.
}

