% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{set_due_date}
\alias{set_due_date}
\title{Calls corresponding RTM task method}
\usage{
set_due_date(task, due)
}
\arguments{
\item{task}{the task}

\item{due}{due date}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

