% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{set_location}
\alias{set_location}
\title{Calls corresponding RTM task method}
\usage{
set_location(task, location_id)
}
\arguments{
\item{task}{the task}

\item{location_id}{location id}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

