% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_funs.r
\name{set_name}
\alias{set_name}
\title{Calls corresponding RTM task method}
\usage{
set_name(task, name)
}
\arguments{
\item{task}{the task}

\item{name}{the task name}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

