% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/list.r
\name{set_name_list}
\alias{set_name_list}
\title{Perform the list operation}
\usage{
set_name_list(list_id, name)
}
\arguments{
\item{list_id}{the list id}

\item{name}{the list name}
}
\value{
response from RTM server
}
\description{
Perform the list operation
}

