% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{set_priority}
\alias{set_priority}
\title{Calls corresponding RTM task method}
\usage{
set_priority(task, priority)
}
\arguments{
\item{task}{the task}

\item{priority}{priority}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

