% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_funs.r
\name{set_tags}
\alias{set_tags}
\title{Calls corresponding RTM task method}
\usage{
set_tags(task, tags)
}
\arguments{
\item{task}{the task}

\item{tags}{tags}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

