% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{setEstimate}
\alias{setEstimate}
\title{Calls corresponding RTM task method}
\usage{
setEstimate(task, estimate)
}
\arguments{
\item{task}{the task}

\item{estimate}{time estimate}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

