% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{setName}
\alias{setName}
\title{Calls corresponding RTM task method}
\usage{
setName(task, name)
}
\arguments{
\item{task}{the task}

\item{name}{the task name}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

