% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{setPriority}
\alias{setPriority}
\title{Calls corresponding RTM task method}
\usage{
setPriority(task, priority)
}
\arguments{
\item{task}{the task}

\item{priority}{priority}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

