% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{setRecurrence}
\alias{setRecurrence}
\title{Calls corresponding RTM task method}
\usage{
setRecurrence(task, `repeat`)
}
\arguments{
\item{task}{the task}

\item{`repeat`}{recurrence}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

