% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task_funs.r
\name{setURL}
\alias{setURL}
\title{Calls corresponding RTM task method}
\usage{
setURL(task, url)
}
\arguments{
\item{task}{the task}

\item{url}{url}
}
\value{
response from the server
}
\description{
Calls corresponding RTM task method
}

