% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.r
\name{sign_request}
\alias{sign_request}
\title{Sign an API request with a shared secret and MD5 hash according
to the rule at
https://www.rememberthemilk.com/services/api/authentication.rtm}
\usage{
sign_request(query)
}
\arguments{
\item{query}{a list of all query parameters}
}
\value{
a hexadecimal string value
}
\description{
Sign an API request with a shared secret and MD5 hash according
to the rule at
https://www.rememberthemilk.com/services/api/authentication.rtm
}

