% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/response.r
\name{status.rtm_response}
\alias{status.rtm_response}
\title{Get the status of a response}
\usage{
\method{status}{rtm_response}(x)
}
\arguments{
\item{x}{an rtm_response object}
}
\value{
"ok" or "fail"
}
\description{
Get the status of a response
}

