% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{task_method_args}
\alias{task_method_args}
\title{A list of the arguments to RTM task method calls}
\usage{
task_method_args()
}
\value{
the list
}
\description{
A list of the arguments to RTM task method calls
}

