% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmauth_token.r
\docType{class}
\name{Token-class}
\alias{Token-class}
\alias{TokenRTM}
\title{RTMAuth token objects.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
TokenRTM
}
\description{
See httr::Token for more info.
}
\keyword{datasets}

