% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.r
\name{unarchive_list}
\alias{unarchive_list}
\title{Perform the list operation}
\usage{
unarchive_list(list_id)
}
\arguments{
\item{list_id}{the list id}
}
\value{
response from RTM server
}
\description{
Perform the list operation
}

