% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httr-helpers.r
\name{VERB_n}
\alias{VERB_n}
\title{HTTP exponential backoff.}
\usage{
VERB_n(VERB, n = 5)
}
\arguments{
\item{VERB}{original function}

\item{n}{max times}
}
\value{
function
}
\description{
Copied from jennybc's googlesheets on github. Used to avoid rate
limit issues:Let's see if this works best, or if directly limiting
to 1/sec or max 3/sec somehow would be more effective.
}
\details{
https://www.rememberthemilk.com/services/api/ratelimit.rtm
}

