% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.r
\name{verify_token}
\alias{verify_token}
\title{Checks with the API to see whether the user's PAT is still valid.}
\usage{
verify_token()
}
\value{
logical indicating whether the token is still valid or not
}
\description{
from the website: "\code{auth_token}'s can and do expire
(for example, if the user revokes the permissions they granted to
your application)"
}

