% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/task.r
\name{filter_task_info}
\alias{filter_task_info}
\title{Process task info from rtm_search.}
\usage{
filter_task_info(df)
}
\arguments{
\item{df}{a data frame of tasks returned by rtm_search}
}
\description{
Process task info from rtm_search.
}

