% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtm.R
\name{plot.rtm}
\alias{plot.rtm}
\title{Plot the risk terrain map}
\usage{
\method{plot}{rtm}(x, what = c("relrisk", "prediction", "outcome_count"), ...)
}
\arguments{
\item{x}{An \code{rtm} object.}

\item{what}{Cell value to map: \code{"relrisk"} (relative risk score, default),
\code{"prediction"} (expected event count), or \code{"outcome_count"}.}

\item{...}{Passed on to \link[sf:plot_sf]{plot.sf()}.}
}
\description{
Plot the risk terrain map
}
