% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{rtm_search}
\alias{rtm_search}
\title{Search tasks using the RTM advanced search feature.}
\usage{
rtm_search(filter, completed = FALSE)
}
\arguments{
\item{filter}{the search terms passed to RTM advanced search}

\item{completed}{return completed tasks?}
}
\value{
a list of data frames contained in the response from rtm_req
    for an advanced search
}
\description{
See https://www.rememberthemilk.com/help/?ctx=basics.search.advanced
for specific parameters.
}

