% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtmauth_signature.r
\name{rtmauth_signature}
\alias{rtmauth_signature}
\title{Sign parameters for an RTM request.}
\usage{
rtmauth_signature(url, app, token = NULL, frob = NULL, ...)
}
\arguments{
\item{url}{Url of request}

\item{app}{RTM app object}

\item{token}{token string}

\item{frob}{frob string}

\item{...}{additional parameters}
}
\value{
list object of length two containing the signature and api_key
}
\description{
See "Signing Requests" at:
https://www.rememberthemilk.com/services/api/authentication.rtm
}
\details{
Follows the design of httr::oauth_signature.
}

